Починенные тесты для ManagerNewsControllerTest

+++ service-impl/src/main/java/com/vet24/service/news/NewsServiceImpl.java
@@ -113,9 +113,11 @@ public class NewsServiceImpl extends ReadWriteServiceImpl<Long, News> implements
                 .collect(Collectors.toList());
 
         for (News unpublishedNews: newsDao.getNewsById(unpublishNewsIds)) {
-            discordService.deleteMessage(
-                    unpublishedNews.getDiscordMessage()
-                            .getDiscordMsgId());
+            if(unpublishedNews.getDiscordMessage() != null) {
+                discordService.deleteMessage(
+                        unpublishedNews.getDiscordMessage()
+                                .getDiscordMsgId());
+            }
             unpublishedNews.setDiscordMessage(null);
         }
